new gpt.bat file
it sets the -Duser.home to prevent loading module updates and
sets -Dsnap.dataio.reader.tileHeight=%blocksize% -Dsnap.dataio.reader.tileWidth=%blocksize%
which were not effective before. These need to be set as VM parameters and not as program parameters. -J-D is also not working. Only works with the original gpt.exe.